requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2==2.9.9
pydantic==2.5.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
